Add files via upload
This is the code for image extraction and processing the numbers on the plate. We have used the MATLAB , we can also use open cv. Code is very similar to C language. Moreover we have taken random vehicle plates for processing the image. The character resemblance is generally of alphanumeric characters.